@@ -42,42 +42,50 @@
   ],
   "bugs": "https://github.com/salesforcecli/plugin-custom-metadata/issues",
   "dependencies": {
-    "@oclif/config": "^1.18.1",
-    "@salesforce/command": "^4.2.0",
-    "@salesforce/core": "^2.31.0",
-    "@salesforce/ts-types": "^1.2.2",
+    "@oclif/core": "^1.9.0",
+    "@salesforce/command": "^5.1.2",
+    "@salesforce/core": "^3.19.0",
+    "@salesforce/ts-types": "^1.5.20",
+    "fast-xml-parser": "^4.0.7",
     "tslib": "^2"
   },
   "devDependencies": {
-    "@oclif/dev-cli": "^1.26.6",
-    "@oclif/plugin-command-snapshot": "^2.2.2",
-    "@oclif/plugin-help": "^3.2.10",
-    "@oclif/test": "^1.2.8",
-    "@salesforce/dev-config": "^2.1.2",
-    "@salesforce/dev-scripts": "1.0.2",
+    "@oclif/plugin-command-snapshot": "^3.1.3",
+    "@salesforce/cli-plugins-testkit": "^1.5.35",
+    "@salesforce/dev-config": "^3.0.1",
+    "@salesforce/dev-scripts": "^2.0.2",
     "@salesforce/plugin-command-reference": "^1.3.17",
-    "@salesforce/ts-sinon": "^1.1.2",
+    "@salesforce/prettier-config": "^0.0.2",
+    "@salesforce/ts-sinon": "^1.3.21",
     "@semantic-release/git": "^9.0.0",
-    "@types/chai": "^4.2.11",
-    "@types/jsforce": "1.9.35",
-    "@types/mocha": "^7.0.2",
-    "@types/node": "^14.0.9",
-    "@types/xml2js": "^0.4.5",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/parser": "^4.2.0",
     "bluebird": "3.7.2",
     "chai": "^4.2.0",
-    "globby": "^11.0.1",
-    "husky": "^4.2.5",
+    "cz-conventional-changelog": "^3.3.0",
+    "eslint": "^7.27.0",
+    "eslint-config-prettier": "^6.15.0",
+    "eslint-config-salesforce": "^0.1.6",
+    "eslint-config-salesforce-license": "^0.1.6",
+    "eslint-config-salesforce-typescript": "^0.2.8",
+    "eslint-plugin-header": "^3.1.1",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsdoc": "^39.3.2",
+    "eslint-plugin-prettier": "^3.4.1",
+    "husky": "^7.0.4",
+    "jsforce": "2.0.0-beta.10",
     "mocha": "^9.1.3",
     "nyc": "^15.1.0",
-    "semantic-release": "^17.0.8",
+    "oclif": "^3.0.1",
+    "prettier": "^2.6.2",
+    "pretty-quick": "^3.1.0",
     "shx": "0.2.2",
-    "sinon": "9.0.2",
-    "ts-node": "^8.10.2",
-    "typescript": "^4.3.2",
-    "xml2js": "^0.4.23"
+    "sinon": "10.0.0",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.2"
   },
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=14.0.0"
   },
   "files": [
     "/csvtojson",
