@@ -4,24 +4,35 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-
+import * as fs from 'fs';
+import * as path from 'path';
 import { flags, SfdxCommand } from '@salesforce/command';
-import { Messages, fs, SfdxError } from '@salesforce/core';
-import { AnyJson } from '@salesforce/ts-types';
+import { Messages, SfError } from '@salesforce/core';
+import { CustomField } from 'jsforce/api/metadata';
 import { CreateUtil } from '../../../../lib/helpers/createUtil';
-import { FileWriter } from '../../../../lib/helpers/fileWriter';
-import { ValidationUtil } from '../../../../lib/helpers/validationUtil';
+import {
+  validateMetadataRecordName,
+  validateMetadataTypeName,
+  validateLessThanForty,
+} from '../../../../lib/helpers/validationUtil';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
 
 // Load the specific messages for this file. Messages from @salesforce/command, @salesforce/core,
 // or any library that is using the messages framework can also be loaded this way.
-const messages = Messages.loadMessages(
-  '@salesforce/plugin-custom-metadata',
-  'createRecord'
-);
+const messages = Messages.loadMessages('@salesforce/plugin-custom-metadata', 'createRecord');
 
+interface CmdtRecordCreateResponse {
+  typename: string;
+  recordname: string;
+  label: string;
+  inputdir: string;
+  outputdir: string;
+  protectedFlag: boolean;
+  varargs: Record<string, unknown>;
+  fileData: CustomField[];
+}
 export default class Create extends SfdxCommand {
   public static description = messages.getMessage('commandDescription');
   public static longDescription = messages.getMessage('commandLongDescription');
@@ -42,17 +53,21 @@ export default class Create extends SfdxCommand {
       description: messages.getMessage('typenameFlagDescription'),
       longDescription: messages.getMessage('typenameFlagLongDescription'),
       required: true,
+      parse: async (input) => Promise.resolve(validateMetadataTypeName(input)),
     }),
     recordname: flags.string({
       char: 'n',
       description: messages.getMessage('recordNameFlagDescription'),
       longDescription: messages.getMessage('recordNameFlagLongDescription'),
       required: true,
+      parse: async (input) => Promise.resolve(validateMetadataRecordName(input)),
     }),
     label: flags.string({
       char: 'l',
       description: messages.getMessage('labelFlagDescription'),
       longDescription: messages.getMessage('labelFlagLongDescription'),
+      parse: async (input) =>
+        Promise.resolve(validateLessThanForty(input, messages.getMessage('notAValidLabelNameError', [input]))),
     }),
     protected: flags.string({
       char: 'p',
@@ -65,68 +80,44 @@ export default class Create extends SfdxCommand {
       char: 'i',
       description: messages.getMessage('inputDirectoryFlagDescription'),
       longDescription: messages.getMessage('inputDirectoryFlagLongDescription'),
-      default: 'force-app/main/default/objects',
+      default: path.join('force-app', 'main', 'default', 'objects'),
     }),
     outputdir: flags.directory({
       char: 'd',
       description: messages.getMessage('outputDirectoryFlagDescription'),
-      longDescription: messages.getMessage(
-        'outputDirectoryFlagLongDescription'
-      ),
-      default: 'force-app/main/default/customMetadata',
+      longDescription: messages.getMessage('outputDirectoryFlagLongDescription'),
+      default: path.join('force-app', 'main', 'default', 'customMetadata'),
     }),
   };
 
   protected static varargs = {
     required: false,
-    validator: (name, value) => {
+    validator: (name: string): void => {
       // only custom fields allowed
       if (!name.endsWith('__c')) {
         const errMsg = `Invalid parameter [${name}] found`;
         const errName = 'InvalidVarargName';
         const errAction = messages.getMessage('errorInvalidCustomField');
-        throw new SfdxError(errMsg, errName, [errAction]);
+        throw new SfError(errMsg, errName, [errAction]);
       }
     },
   };
 
-  // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
   protected static requiresProject = true;
 
-  public async run(): Promise<AnyJson> {
+  // eslint-disable-next-line @typescript-eslint/member-ordering
+  public async run(): Promise<CmdtRecordCreateResponse> {
     try {
-      const validator = new ValidationUtil();
       const createUtil = new CreateUtil();
-      const fileWriter = new FileWriter();
-      let typename = this.flags.typename;
-      const recordname = this.flags.recordname;
-      const label = this.flags.label || this.flags.recordname;
-      const protectedFlag = this.flags.protected || 'false';
-      const inputdir = this.flags.inputdir || 'force-app/main/default/objects';
-      const outputdir =
-        this.flags.outputdir || 'force-app/main/default/customMetadata';
+      let typename = this.flags.typename as string;
+      const recordname = this.flags.recordname as string;
+      const label = (this.flags.label as string) ?? recordname;
+      const protectedFlag = (this.flags.protected as string) === 'true';
+      const inputdir = this.flags.inputdir as string;
+      const outputdir = this.flags.outputdir as string;
       const dirName = createUtil.appendDirectorySuffix(typename);
-      const fieldDirPath = `${fileWriter.createDir(inputdir)}${dirName}/fields`;
-
-      if (!validator.validateMetadataTypeName(typename)) {
-        throw new SfdxError(
-          messages.getMessage('notValidAPINameError', [typename])
-        );
-      }
-
-      if (!validator.validateMetadataRecordName(recordname)) {
-        throw new SfdxError(
-          messages.getMessage('notAValidRecordNameError', [recordname])
-        );
-      }
-
-      if (!validator.validateLessThanForty(label)) {
-        throw new SfdxError(
-          messages.getMessage('notAValidLabelNameError', [label])
-        );
-      }
-
-      const fileNames = await fs.readdir(fieldDirPath);
+      const fieldDirPath = path.join(inputdir, dirName, 'fields');
+      const fileNames = await fs.promises.readdir(fieldDirPath);
 
       // forgive them if they passed in type__mdt, and cut off the __mdt
       if (typename.endsWith('__mdt')) {
@@ -134,7 +125,7 @@ export default class Create extends SfdxCommand {
       }
 
       // if customMetadata folder does not exist, create it
-      await fs.mkdirp(outputdir);
+      await fs.promises.mkdir(outputdir, { recursive: true });
 
       const fileData = await createUtil.getFileData(fieldDirPath, fileNames);
 
@@ -149,15 +140,7 @@ export default class Create extends SfdxCommand {
         fileData,
       });
 
-      this.ux.log(
-        messages.getMessage('successResponse', [
-          typename,
-          recordname,
-          label,
-          protectedFlag,
-          outputdir,
-        ])
-      );
+      this.ux.log(messages.getMessage('successResponse', [typename, recordname, label, protectedFlag, outputdir]));
 
       // Return an object to be displayed with --json
       return {
@@ -171,7 +154,9 @@ export default class Create extends SfdxCommand {
         fileData,
       };
     } catch (err) {
-      this.ux.log(err.message);
+      if (err instanceof Error) {
+        this.ux.log(err.message);
+      }
     }
   }
 }
