@@ -4,35 +4,35 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-
-import { fs } from '@salesforce/core';
-import { parseString } from 'xml2js';
+import * as fs from 'fs';
+import * as path from 'path';
+import { CustomField } from 'jsforce/api/metadata';
+import { XMLParser } from 'fast-xml-parser';
 import { CreateConfig } from '../interfaces/createConfig';
-import { CustomField } from '../interfaces/customField';
 import { Templates } from '../templates/templates';
 
+interface CustomFieldFile {
+  CustomField: CustomField;
+}
+
+const fieldTypeMap = {
+  Checkbox: 'boolean',
+  Date: 'date',
+  DateTime: 'dateTime',
+  Email: 'string',
+  Phone: 'string',
+  Picklist: 'string',
+  Text: 'string',
+  TextArea: 'string',
+  LongTextArea: 'string',
+  Url: 'string',
+};
 // NOTE: the template string indentation is important to output well-formatted XML. Altering that whitespace will change the whitespace of the output.
 export class CreateUtil {
-  private fieldTypeMap: object;
-
   /**
    * Number and Percent types will be int or double depending on their respective scale values.
    * If the scale === 0, it is an int, otherwise it is a double
    */
-  constructor() {
-    this.fieldTypeMap = {
-      Checkbox: 'boolean',
-      Date: 'date',
-      DateTime: 'dateTime',
-      Email: 'string',
-      Phone: 'string',
-      Picklist: 'string',
-      Text: 'string',
-      TextArea: 'string',
-      LongTextArea: 'string',
-      Url: 'string',
-    };
-  }
 
   /**
    * Creates the Custom Metadata Record
@@ -40,38 +40,36 @@ export class CreateUtil {
    * @param  createConfig Properties include typename, recname, label, protection, varargs, and fileData
    * @return void
    */
-  public async createRecord(createConfig: CreateConfig) {
-    const outputFilePath = `${createConfig.outputdir}/${createConfig.typename}.${createConfig.recordname}.md-meta.xml`;
+  public async createRecord(createConfig: CreateConfig): Promise<void> {
+    const outputFilePath = path.join(
+      createConfig.outputdir,
+      `${createConfig.typename}.${createConfig.recordname}.md-meta.xml`
+    );
     const newRecordContent = this.getRecordTemplate(
       createConfig.label,
       createConfig.protected,
-      this.buildCustomFieldXml(
-        createConfig.fileData,
-        createConfig.varargs,
-        createConfig.ignorefields
-      )
+      this.buildCustomFieldXml(createConfig.fileData, createConfig.varargs, createConfig.ignorefields)
     );
 
-    return fs.writeFile(outputFilePath, newRecordContent);
+    return fs.promises.writeFile(outputFilePath, newRecordContent);
   }
 
-  public async getFileData(fieldDirPath, fileNames) {
-    const ret = [];
-    let filePath = '';
-    let fileData;
-    let str = '';
-
-    for (const file of fileNames) {
-      filePath = `${fieldDirPath}/${file}`;
-      fileData = await fs.readFile(filePath);
-      str = fileData.toString('utf8');
-
-      parseString(str, (err, res) => {
-        ret.push(res);
-      });
-    }
-
-    return ret;
+  /**
+   *
+   * @param fieldDirPath path to a /fields folder that contains all the fields to read
+   * @param fileNames filenames in that folder that should be read
+   * @returns CustomField[]
+   */
+  public async getFileData(fieldDirPath: string, fileNames: string[]): Promise<CustomField[]> {
+    const parser = new XMLParser();
+    return Promise.all(
+      fileNames
+        .map((file) => path.join(fieldDirPath, file))
+        .map(async (filePath) => {
+          const fileData = await fs.promises.readFile(filePath, 'utf8');
+          return (parser.parse(fileData) as CustomFieldFile).CustomField;
+        })
+    );
   }
 
   /**
@@ -79,12 +77,8 @@ export class CreateUtil {
    *
    * @param  typename Name of file
    */
-  public appendDirectorySuffix(typename: string) {
-    if (typename.endsWith('__mdt')) {
-      return typename;
-    } else {
-      return typename + '__mdt';
-    }
+  public appendDirectorySuffix(typename: string): string {
+    return typename.endsWith('__mdt') ? typename : `${typename}__mdt`;
   }
 
   /**
@@ -94,27 +88,11 @@ export class CreateUtil {
    * @param  fieldName Name of the field
    * @return {string} Type used by a custom metadata record
    */
-  public getFieldPrimitiveType(fileData: CustomField[] = [], fieldName = '') {
-    let thisFieldName = '';
-    let type = '';
-    let ret = 'string';
-
-    for (const file of fileData) {
-      thisFieldName = file.CustomField.fullName[0];
-      type = file.CustomField.type[0];
-
-      if (type === 'Number' || type === 'Percent') {
-        ret = this.getNumberType(type, file.CustomField.scale[0]);
-      } else {
-        ret = this.fieldTypeMap[type];
-      }
-
-      if (thisFieldName === fieldName) {
-        return ret;
-      }
-    }
-
-    return ret;
+  public getFieldPrimitiveType(fileData: CustomField[] = [], fieldName?: string): string {
+    const matchingFile = fileData.find((file) => file.fullName === fieldName);
+    return matchingFile && ['Number', 'Percent'].includes(matchingFile.type)
+      ? this.getNumberType(matchingFile.type, matchingFile.scale)
+      : (fieldTypeMap[matchingFile?.type] as string) ?? 'string';
   }
 
   /**
@@ -124,19 +102,10 @@ export class CreateUtil {
    * @param  fieldName Name of the field
    * @return {string} Data Type of the field.
    */
-  public getFieldDataType(fileData: CustomField[] = [], fieldName = '') {
-    let thisFieldName = '';
-    let type = '';
-
-    for (const file of fileData) {
-      thisFieldName = file.CustomField.fullName[0];
-      type = file.CustomField.type[0];
-
-      if (thisFieldName === fieldName) {
-        return type;
-      }
-    }
+  public getFieldDataType(fileData: CustomField[] = [], fieldName = ''): CustomField['type'] {
+    return fileData.find((file) => file.fullName === fieldName)?.type;
   }
+
   /**
    * Goes through the file data that has been genreated and gets all of the field names and adds the
    * name of the field that is used as the label for metadata record
@@ -145,15 +114,8 @@ export class CreateUtil {
    * @param  nameField name of the column that is going to be used for the name of the metadata record
    * @return [] Array of field names
    */
-  public getFieldNames(fileData, nameField) {
-    const metadataTypeFields = [];
-    for (const file of fileData) {
-      const fullName: string = file.CustomField.fullName[0];
-      metadataTypeFields.push(fullName);
-    }
-    metadataTypeFields.push(nameField);
-
-    return metadataTypeFields;
+  public getFieldNames(fileData: CustomField[], nameField: string): string[] {
+    return [...fileData.map((file) => file.fullName), nameField];
   }
 
   /**
@@ -166,23 +128,17 @@ export class CreateUtil {
    */
   private buildCustomFieldXml(
     fileData: CustomField[],
-    cliParams: object,
+    cliParams: Record<string, string>,
     ignoreFields: boolean
-  ) {
+  ): string {
     let ret = '';
-    let type = '';
-    let dataType = '';
     const templates = new Templates();
     for (const fieldName of Object.keys(cliParams)) {
-      type = this.getFieldPrimitiveType(fileData, fieldName);
-      dataType = this.getFieldDataType(fileData, fieldName);
+      const type = this.getFieldPrimitiveType(fileData, fieldName);
+      const dataType = this.getFieldDataType(fileData, fieldName);
       // Added functionality to handle the igonre fields scenario.
-      if (templates.canConvert(dataType)) {
+      if (templates.canConvert(dataType) || !ignoreFields) {
         ret += this.getFieldTemplate(fieldName, cliParams[fieldName], type);
-      } else {
-        if (!ignoreFields) {
-          ret += this.getFieldTemplate(fieldName, cliParams[fieldName], type);
-        }
       }
     }
 
@@ -197,15 +153,8 @@ export class CreateUtil {
    * @param  scale 0 or another number
    * @return {string} int or double
    */
-  private getNumberType(type: string, scale: string) {
-    if (
-      (type === 'Number' && parseFloat(scale) === 0) ||
-      (type === 'Percent' && parseFloat(scale) === 0)
-    ) {
-      return 'int';
-    }
-
-    return 'double';
+  private getNumberType(type: string, scale: number): 'int' | 'double' {
+    return ['Number', 'Percent'].includes(type) && scale === 0 ? 'int' : 'double';
   }
 
   /**
@@ -216,18 +165,11 @@ export class CreateUtil {
    * @param  type Field type (i.e. boolean, dateTime, date, string, double)
    * @return {string} String representation of XML
    */
-  private getFieldTemplate(fieldName: string, val: string, type: string) {
-    // update to handle the null situation
-    let value = '';
-    const cleanValue = String(val)
-      .replace(/&/g, '&amp;')
-      .replace(/</g, '&lt;')
-      .replace(/>/g, '&gt;');
-    if (val === null || val === '') {
-      value = '<value xsi:nil="true"/>';
-    } else {
-      value = `<value xsi:type="xsd:${type}">${cleanValue}</value>`;
-    }
+  private getFieldTemplate(fieldName: string, val: string, type: string): string {
+    const cleanValue = String(val).replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;');
+    const value =
+      val === null || val === '' ? '<value xsi:nil="true"/>' : `<value xsi:type="xsd:${type}">${cleanValue}</value>`;
+
     return `
     <values>
         <field>${fieldName}</field>
@@ -243,11 +185,7 @@ export class CreateUtil {
    * @param  values Template string representation of values
    * @return {string} String representation of XML
    */
-  private getRecordTemplate(
-    label: string,
-    protection: boolean,
-    values: string
-  ) {
+  private getRecordTemplate(label: string, protection: boolean, values: string): string {
     return `
 <?xml version="1.0" encoding="UTF-8"?>
 <CustomMetadata xmlns="http://soap.sforce.com/2006/04/metadata" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
