@@ -1,333 +1,377 @@
 /*
- * Copyright (c) 2018-2020, salesforce.com, inc.
+ * Copyright (c) 2020, salesforce.com, inc.
  * All rights reserved.
- * SPDX-License-Identifier: BSD-3-Clause
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
 
 import { expect } from '@salesforce/command/lib/test';
 import { Templates } from '../../src/lib/templates/templates';
 
 describe('Templates', () => {
-    const templates = new Templates();
-    describe('createObjectXML', () => {
-        const testData = {
-            "label": "Sample",
-            "pluralLabel": "Samples"
-        };
-        let xml: string = templates.createObjectXML(testData, 'Public');
-        expect(xml.includes(`<label>Sample</label>`)).to.be.true;
-        expect(xml.includes(`<pluralLabel>Samples</pluralLabel>`)).to.be.true;
-        expect(xml.includes(`<visibility>Public</visibility>`)).to.be.true;
+  const templates = new Templates();
+  describe('createObjectXML', () => {
+    const testData = {
+      label: 'Sample',
+      pluralLabel: 'Samples',
+    };
+    const xml: string = templates.createObjectXML(testData, 'Public');
+    expect(xml).to.include('<label>Sample</label>');
+    expect(xml).to.include('<pluralLabel>Samples</pluralLabel>');
+    expect(xml).to.include('<visibility>Public</visibility>');
+  });
+
+  describe('createFieldXML', () => {
+    it('Checkbox field created', () => {
+      const data: any = { fullName: 'Checkbox', defaultValue: 'false', type: 'Checkbox', label: 'test' };
+      const xml: string = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>Checkbox__c</fullName>');
+      expect(xml).to.include('<type>Checkbox</type>');
+      expect(xml).to.include('<label>test</label>');
+    });
+    it('Date field created', () => {
+      const data: any = { fullName: 'Date', type: 'Date', label: 'test' };
+      const xml: string = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>Date__c</fullName>');
+      expect(xml).to.include('<type>Date</type>');
+      expect(xml).to.include('<label>test</label>');
+    });
+    it('DateTime field created', () => {
+      const data: any = { fullName: 'DateTime', type: 'DateTime', label: 'test' };
+      const xml: string = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>DateTime__c</fullName>');
+      expect(xml).to.include('<type>DateTime</type>');
+      expect(xml).to.include('<label>test</label>');
+    });
+    it('Email field created', () => {
+      const data: any = { fullName: 'Email', type: 'Email', label: 'test', unique: false };
+      const xml: string = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>Email__c</fullName>');
+      expect(xml).to.include('<type>Email</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<unique>false</unique>');
+    });
+    it('Number field created', () => {
+      const data: any = {
+        fullName: 'Number',
+        type: 'Number',
+        externalId: 'true',
+        label: 'test',
+        precision: '18',
+        scale: '0',
+        unique: false,
+      };
+      const xml: string = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>Number__c</fullName>');
+      expect(xml).to.include('<type>Number</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<unique>false</unique>');
+      expect(xml).to.include('<precision>18</precision>');
+      expect(xml).to.include('<scale>0</scale>');
+      expect(xml).to.include('<externalId>true</externalId>');
+    });
+    it('Percent field created', () => {
+      const data: any = { fullName: 'Percent', type: 'Percent', label: 'test', precision: '18', scale: '0' };
+      const xml: string = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>Percent__c</fullName>');
+      expect(xml).to.include('<type>Percent</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<precision>18</precision>');
+      expect(xml).to.include('<scale>0</scale>');
+    });
+    it('Phone field created', () => {
+      const data: any = { fullName: 'Phone', type: 'Phone', label: 'test' };
+      const xml: string = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>Phone__c</fullName>');
+      expect(xml).to.include('<type>Phone</type>');
+      expect(xml).to.include('<label>test</label>');
+    });
+    it('Picklist field created', () => {
+      const data: any = {
+        fullName: 'Picklist',
+        type: 'Picklist',
+        label: 'test',
+        valueSet: {
+          restricted: 'true',
+          valueSetDefinition: {
+            sorted: 'false',
+            value: [
+              { fullName: 'value', label: 'value' },
+              { fullName: 'value2', label: 'value2' },
+            ],
+          },
+        },
+      };
+      const xml: string = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>Picklist__c</fullName>');
+      expect(xml).to.include('<type>Picklist</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<fullName>value</fullName>');
+      expect(xml).to.include('<fullName>value2</fullName>');
+      expect(xml).to.include('<label>value</label>');
+      expect(xml).to.include('<label>value2</label>');
+    });
+    it('should not add value sets for a MultiselectPicklist field', () => {
+      const data: any = {
+        fullName: 'Picklist',
+        type: 'MultiselectPicklist',
+        label: 'test',
+        valueSet: {
+          restricted: 'true',
+          valueSetDefinition: {
+            sorted: 'false',
+            value: [
+              { fullName: 'value', label: 'value' },
+              { fullName: 'value2', label: 'value2' },
+            ],
+          },
+        },
+      };
+      const xml = templates.createFieldXML(data, true);
+      expect(xml).to.include('<fullName>Picklist__c</fullName>');
+      expect(xml).to.include('<type>LongTextArea</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.not.include('<fullName>value</fullName>');
+      expect(xml).to.not.include('<fullName>value2</fullName>');
+      expect(xml).to.not.include('<label>value</label>');
+      expect(xml).to.not.include('<label>value2</label>');
+    });
+    it('Text field created', () => {
+      const data: any = { fullName: 'Text', type: 'Text', label: 'test', unique: false, length: '100' };
+      const xml = templates.createFieldXML(data, false);
+      expect(xml);
+      expect(xml).to.include('<fullName>Text__c</fullName>');
+      expect(xml).to.include('<type>Text</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<unique>false</unique>');
+      expect(xml).to.include('<length>100</length>');
+    });
+    it('default to Text field', () => {
+      const data: any = { fullName: 'Text', type: 'Lookup', label: 'test', unique: false, length: '100' };
+      const xml = templates.createFieldXML(data, true);
+      expect(xml).to.include('<fullName>Text__c</fullName>');
+      expect(xml).to.include('<type>Text</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<unique>false</unique>');
+      expect(xml).to.include('<length>100</length>');
+    });
+    it('TextArea field created', () => {
+      const data: any = { fullName: 'TextArea', type: 'TextArea', label: 'test', unique: false, length: '100' };
+      const xml = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>TextArea__c</fullName>');
+      expect(xml).to.include('<type>TextArea</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<unique>false</unique>');
+      expect(xml).to.include('<length>100</length>');
+    });
+    it('LongTextArea field created', () => {
+      const data: any = {
+        fullName: 'LongTextArea',
+        type: 'LongTextArea',
+        label: 'test',
+        length: '32768',
+        visibleLines: 3,
+      };
+      const xml: string = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>LongTextArea__c</fullName>');
+      expect(xml).to.include('<type>LongTextArea</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<length>32768</length>');
+      expect(xml).to.include('<visibleLines>3</visibleLines>');
+    });
+    it('default to LongTextArea field', () => {
+      const data: any = { fullName: 'LongTextArea', type: 'Html', label: 'test', length: '32768', visibleLines: 3 };
+      const xml = templates.createFieldXML(data, true);
+      expect(xml).to.include('<fullName>LongTextArea__c</fullName>');
+      expect(xml).to.include('<type>LongTextArea</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<length>32768</length>');
+      expect(xml).to.include('<visibleLines>3</visibleLines>');
     });
 
-    describe('createFieldXML', () => {
-
-        it('Checkbox field created', () => {
-            let data: any = { fullName: 'Checkbox', defaultValue: 'false', type: 'Checkbox', label: 'test' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>Checkbox__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Checkbox</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-        });
-        it('Date field created', () => {
-            let data: any = { fullName: 'Date', type: 'Date', label: 'test' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>Date__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Date</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-        });
-        it('DateTime field created', () => {
-            let data: any = { fullName: 'DateTime', type: 'DateTime', label: 'test' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>DateTime__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>DateTime</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-        });
-        it('Email field created', () => {
-            let data: any = { fullName: 'Email', type: 'Email', label: 'test', unique: 'false' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>Email__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Email</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<unique>false</unique>`)).to.be.true;
-        });
-        it('Number field created', () => {
-            let data: any = { fullName: 'Number', type: 'Number', externalId: 'true', label: 'test', precision: '18', scale: '0', unique: 'false' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>Number__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Number</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<unique>false</unique>`)).to.be.true;
-            expect(xml.includes(`<precision>18</precision>`)).to.be.true;
-            expect(xml.includes(`<scale>0</scale>`)).to.be.true;
-            expect(xml.includes(`<externalId>true</externalId>`)).to.be.true;
-        });
-        it('Percent field created', () => {
-            let data: any = { fullName: 'Percent', type: 'Percent', label: 'test', precision: '18', scale: '0' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>Percent__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Percent</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<precision>18</precision>`)).to.be.true;
-            expect(xml.includes(`<scale>0</scale>`)).to.be.true;
-        });
-        it('Phone field created', () => {
-            let data: any = { fullName: 'Phone', type: 'Phone', label: 'test' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>Phone__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Phone</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-        });
-        it('Picklist field created', () => {
-            let data: any = {
-                fullName: 'Picklist', type: 'Picklist', label: 'test', valueSet: {
-                    restricted: 'true',
-                    valueSetDefinition: {
-                        sorted: 'false',
-                        value: [{ fullName: 'value', label: 'value' }, { fullName: 'value2', label: 'value2' }]
-                    }
-                }
-            };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>Picklist__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Picklist</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<fullName>value</fullName>`)).to.be.true;
-            expect(xml.includes(`<fullName>value2</fullName>`)).to.be.true;
-            expect(xml.includes(`<label>value</label>`)).to.be.true;
-            expect(xml.includes(`<label>value2</label>`)).to.be.true;
-        });
-        it('should not add value sets for a MultiselectPicklist field', () => {
-            let data: any = {
-                fullName: 'Picklist', type: 'MultiselectPicklist', label: 'test', valueSet: {
-                    restricted: 'true',
-                    valueSetDefinition: {
-                        sorted: 'false',
-                        value: [{ fullName: 'value', label: 'value' }, { fullName: 'value2', label: 'value2' }]
-                    }
-                }
-            };
-            let xml: String = templates.createFieldXML(data, true);
-            expect(xml.includes(`<fullName>Picklist__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>LongTextArea</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<fullName>value</fullName>`)).to.be.false;
-            expect(xml.includes(`<fullName>value2</fullName>`)).to.be.false;
-            expect(xml.includes(`<label>value</label>`)).to.be.false;
-            expect(xml.includes(`<label>value2</label>`)).to.be.false;
-        });
-        it('Text field created', () => {
-            let data: any = { fullName: 'Text', type: 'Text', label: 'test', unique: 'false', length: '100' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>Text__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Text</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<unique>false</unique>`)).to.be.true;
-            expect(xml.includes(`<length>100</length>`)).to.be.true;
-        });
-        it('default to Text field', () => {
-            let data: any = { fullName: 'Text', type: 'Lookup', label: 'test', unique: 'false', length: '100' };
-            let xml: String = templates.createFieldXML(data, true);
-            expect(xml.includes(`<fullName>Text__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Text</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<unique>false</unique>`)).to.be.true;
-            expect(xml.includes(`<length>100</length>`)).to.be.true;
-        });
-        it('TextArea field created', () => {
-            let data: any = { fullName: 'TextArea', type: 'TextArea', label: 'test', unique: 'false', length: '100' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>TextArea__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>TextArea</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<unique>false</unique>`)).to.be.true;
-            expect(xml.includes(`<length>100</length>`)).to.be.true;
-        });
-        it('LongTextArea field created', () => {
-            let data: any = { fullName: 'LongTextArea', type: 'LongTextArea', label: 'test', length: '32768', visibleLines: 3 };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>LongTextArea__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>LongTextArea</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<length>32768</length>`)).to.be.true;
-            expect(xml.includes(`<visibleLines>3</visibleLines>`)).to.be.true;
-        });
-        it('default to LongTextArea field', () => {
-            let data: any = { fullName: 'LongTextArea', type: 'Html', label: 'test', length: '32768', visibleLines: 3 };
-            let xml: String = templates.createFieldXML(data, true);
-            expect(xml.includes(`<fullName>LongTextArea__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>LongTextArea</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<length>32768</length>`)).to.be.true;
-            expect(xml.includes(`<visibleLines>3</visibleLines>`)).to.be.true;
-        });
-        
-        it('Url field created', () => {
-            let data: any = { fullName: 'Url', type: 'Url', label: 'test' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>Url__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Url</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-        });
-        it('default to LongTextArea field', () => {
-            let data: any = { fullName: 'LongTextArea', type: 'Html', label: 'test', length: '32768', visibleLines: 3 };
-            try {
-                templates.createFieldXML(data, false);
-            } catch( ex){
-                expect(ex.message).to.contain('\'Html\' is not a valid field type.');
-            }
-        });
-        it('default to Text field for Currency field', () => {
-            let data: any = { fullName: 'Currency', type: 'Currency', label: 'test', precision: '18', scale: '0', unique: 'false', defaultValue: '1000' };
-            let xml: String = templates.createFieldXML(data, true);
-            expect(xml.includes(`<fullName>Currency__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Text</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<unique>false</unique>`)).to.be.true;
-            expect(xml.includes(`<length>100</length>`)).to.be.true;
-            expect(xml.includes(`<defaultValue>'1000'</defaultValue>`)).to.be.true;
-        });
-        it('should convert Formula fields to long text area type', () => {
-            let data: any = { fullName: 'FormulaText', type: 'Text', label: 'test', externalId: 'false', unique: 'false', inlineHelpText: 'Formula text field' };
-            let xml: String = templates.createFieldXML(data, true);
-            expect(xml.includes(`<fullName>FormulaText__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>LongTextArea</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<length>32768</length>`)).to.be.true;
-        });
-        it('Formula Checkbox field created', () => {
-            let data: any = { fullName: 'FormulaCheckbox', type: 'Checkbox', label: 'test' };
-            let xml: String = templates.createFieldXML(data, false);
-            expect(xml.includes(`<fullName>FormulaCheckbox__c</fullName>`)).to.be.true;
-            expect(xml.includes(`<type>Checkbox</type>`)).to.be.true;
-            expect(xml.includes(`<label>test</label>`)).to.be.true;
-            expect(xml.includes(`<defaultValue>false</defaultValue>`)).to.be.true;
-            expect(xml.includes(`<inlineHelpText>Formula text field</inlineHelpText>`)).to.be.false;
-        });
-
+    it('Url field created', () => {
+      const data: any = { fullName: 'Url', type: 'Url', label: 'test' };
+      const xml = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>Url__c</fullName>');
+      expect(xml).to.include('<type>Url</type>');
+      expect(xml).to.include('<label>test</label>');
     });
-    describe('createDefaultTypeStructure', () => {
-        let struct: any;
-        it('should return Checkbox structure', () => {
-            struct = templates.createDefaultTypeStructure('Checkbox', 'Checkbox', 'test', null);
-            expect(struct.fullName === 'Checkbox').to.be.true;
-            expect(struct.type === 'Checkbox').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-            expect(struct.defaultValue === 'false').to.be.true;
-        });
-        it('should return Date structure', () => {
-            struct = templates.createDefaultTypeStructure('Date', 'Date', 'test', null);
-            expect(struct.fullName === 'Date').to.be.true;
-            expect(struct.type === 'Date').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-        });
-        it('should return DateTime structure', () => {
-            struct = templates.createDefaultTypeStructure('DateTime', 'DateTime', 'test', null);
-            expect(struct.fullName === 'DateTime').to.be.true;
-            expect(struct.type === 'DateTime').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-        });
-        it('should return Email structure', () => {
-            struct = templates.createDefaultTypeStructure('Email', 'Email', 'test', null);
-            expect(struct.fullName === 'Email').to.be.true;
-            expect(struct.type === 'Email').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-            expect(struct.unique === 'false').to.be.true;
-        });
-        it('should return Number structure', () => {
-            struct = templates.createDefaultTypeStructure('Number', 'Number', 'test', null);
-            expect(struct.fullName === 'Number').to.be.true;
-            expect(struct.type === 'Number').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-            expect(struct.precision === 18).to.be.true;
-            expect(struct.scale === 0).to.be.true;
-            expect(struct.unique === 'false').to.be.true;
-        });
-        it('should return Percent structure', () => {
-            struct = templates.createDefaultTypeStructure('Percent', 'Percent', 'test', null);
-            expect(struct.fullName === 'Percent').to.be.true;
-            expect(struct.type === 'Percent').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-            expect(struct.precision === 18).to.be.true;
-            expect(struct.scale === 0).to.be.true;
-        });
-        it('should return Phone structure', () => {
-            struct = templates.createDefaultTypeStructure('Phone', 'Phone', 'test', null);
-            expect(struct.fullName === 'Phone').to.be.true;
-            expect(struct.type === 'Phone').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-        });
-        it('should return Picklist structure', () => {
-            struct = templates.createDefaultTypeStructure('Picklist', 'Picklist', 'test', ['value1', 'value2']);
-            expect(struct.fullName === 'Picklist').to.be.true;
-            expect(struct.type === 'Picklist').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-            expect(struct.valueSet.restricted === 'true').to.be.true;
-            expect(struct.valueSet.valueSetDefinition.sorted === 'false').to.be.true;
-            expect(struct.valueSet.valueSetDefinition.value[0].label === 'value1').to.be.true;
-        });
-        it('should return Phone structure', () => {
-            struct = templates.createDefaultTypeStructure('Phone', 'Phone', 'test', null);
-            expect(struct.fullName === 'Phone').to.be.true;
-            expect(struct.type === 'Phone').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-        });
-        it('should return Text structure', () => {
-            struct = templates.createDefaultTypeStructure('Text', 'Text', 'test', null);
-            expect(struct.fullName === 'Text').to.be.true;
-            expect(struct.type === 'Text').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-            expect(struct.unique === 'false').to.be.true;
-            expect(struct.length === '100').to.be.true;
-        });
-        it('should return TextArea structure', () => {
-            struct = templates.createDefaultTypeStructure('TextArea', 'TextArea', 'test', null);
-            expect(struct.fullName === 'TextArea').to.be.true;
-            expect(struct.type === 'TextArea').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-        });
-        it('should return LongTextArea structure', () => {
-            struct = templates.createDefaultTypeStructure('LongTextArea', 'LongTextArea', 'test', null);
-            expect(struct.fullName === 'LongTextArea').to.be.true;
-            expect(struct.type === 'LongTextArea').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-            expect(struct.length === '32768').to.be.true;
-            expect(struct.visibleLines === 3).to.be.true;
-        });
-        it('should return Url structure', () => {
-            struct = templates.createDefaultTypeStructure('Url', 'Url', 'test', null);
-            expect(struct.fullName === 'Url').to.be.true;
-            expect(struct.type === 'Url').to.be.true;
-            expect(struct.label === 'test').to.be.true;
-        });
-        it('should return Default structure', () => {
-            struct = templates.createDefaultTypeStructure('DefaultA', 'DefaultB', 'DefaultC', null);
-            expect(struct.fullName === 'DefaultA').to.be.true;
-            expect(struct.type === 'DefaultB').to.be.true;
-            expect(struct.label === 'DefaultC').to.be.true;
-        });
+    it('default to LongTextArea field', () => {
+      const data: any = { fullName: 'LongTextArea', type: 'Html', label: 'test', length: '32768', visibleLines: 3 };
+      try {
+        templates.createFieldXML(data, false);
+      } catch (ex) {
+        expect(ex.message).to.contain("'Html' is not a valid field type.");
+      }
     });
-    describe('canConvert', () => {
-        it('type should convert', () => {
-            ['Checkbox',
-                'Date',
-                'DateTime',
-                'Email',
-                'Number',
-                'Percent',
-                'Phone',
-                'Picklist',
-                'Text',
-                'TextArea',
-                'LongTextArea',
-                'Url']
-                .forEach(type => {
-                    expect(templates.canConvert(type)).to.be.true;
-                })
-        });
-        it('type should not convert', () => {
-            expect(templates.canConvert('test')).to.be.false;
-        })
+    it('default to Text field for Currency field', () => {
+      const data: any = {
+        fullName: 'Currency',
+        type: 'Currency',
+        label: 'test',
+        precision: '18',
+        scale: '0',
+        unique: false,
+        defaultValue: '1000',
+      };
+      const xml: string = templates.createFieldXML(data, true);
+      expect(xml).to.include('<fullName>Currency__c</fullName>');
+      expect(xml).to.include('<type>Text</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<unique>false</unique>');
+      expect(xml).to.include('<length>100</length>');
+      expect(xml).to.include("<defaultValue>'1000'</defaultValue>");
     });
+    it('should convert Formula fields to long text area type', () => {
+      const data: any = {
+        fullName: 'FormulaText',
+        type: 'Text',
+        label: 'test',
+        externalId: 'false',
+        unique: false,
+        inlineHelpText: 'Formula text field',
+      };
+      const xml: string = templates.createFieldXML(data, true);
+      expect(xml).to.include('<fullName>FormulaText__c</fullName>');
+      expect(xml).to.include('<type>LongTextArea</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<length>32768</length>');
+    });
+    it('Formula Checkbox field created', () => {
+      const data: any = { fullName: 'FormulaCheckbox', type: 'Checkbox', label: 'test' };
+      const xml: string = templates.createFieldXML(data, false);
+      expect(xml).to.include('<fullName>FormulaCheckbox__c</fullName>');
+      expect(xml).to.include('<type>Checkbox</type>');
+      expect(xml).to.include('<label>test</label>');
+      expect(xml).to.include('<defaultValue>false</defaultValue>');
+      expect(xml.includes('<inlineHelpText>Formula text field</inlineHelpText>')).to.be.false;
+    });
+  });
+  describe('createDefaultTypeStructure', () => {
+    let struct: any;
+    it('should return Checkbox structure', () => {
+      struct = templates.createDefaultTypeStructure('Checkbox', 'Checkbox', 'test', null);
+      expect(struct.fullName).to.equal('Checkbox');
+      expect(struct.type).to.equal('Checkbox');
+      expect(struct.label).to.equal('test');
+      expect(struct.defaultValue).to.equal('false');
+    });
+    it('should return Date structure', () => {
+      struct = templates.createDefaultTypeStructure('Date', 'Date', 'test', null);
+      expect(struct.fullName).to.equal('Date');
+      expect(struct.type).to.equal('Date');
+      expect(struct.label).to.equal('test');
+    });
+    it('should return DateTime structure', () => {
+      struct = templates.createDefaultTypeStructure('DateTime', 'DateTime', 'test', null);
+      expect(struct.fullName).to.equal('DateTime');
+      expect(struct.type).to.equal('DateTime');
+      expect(struct.label).to.equal('test');
+    });
+    it('should return Email structure', () => {
+      struct = templates.createDefaultTypeStructure('Email', 'Email', 'test', null);
+      expect(struct.fullName).to.equal('Email');
+      expect(struct.type).to.equal('Email');
+      expect(struct.label).to.equal('test');
+      expect(struct.unique).to.equal(false);
+    });
+    it('should return Number structure', () => {
+      struct = templates.createDefaultTypeStructure('Number', 'Number', 'test', null);
+      expect(struct.fullName).to.equal('Number');
+      expect(struct.type).to.equal('Number');
+      expect(struct.label).to.equal('test');
+      expect(struct.precision).to.equal(18);
+      expect(struct.scale).to.equal(0);
+      expect(struct.unique).to.equal(false);
+    });
+    it('should return Percent structure', () => {
+      struct = templates.createDefaultTypeStructure('Percent', 'Percent', 'test', null);
+      expect(struct.fullName).to.equal('Percent');
+      expect(struct.type).to.equal('Percent');
+      expect(struct.label).to.equal('test');
+      expect(struct.precision).to.equal(18);
+      expect(struct.scale).to.equal(0);
+    });
+    it('should return Phone structure', () => {
+      struct = templates.createDefaultTypeStructure('Phone', 'Phone', 'test', null);
+      expect(struct.fullName).to.equal('Phone');
+      expect(struct.type).to.equal('Phone');
+      expect(struct.label).to.equal('test');
+    });
+    it('should return Picklist structure', () => {
+      struct = templates.createDefaultTypeStructure('Picklist', 'Picklist', 'test', ['value1', 'value2']);
+      expect(struct.fullName).to.equal('Picklist');
+      expect(struct.type).to.equal('Picklist');
+      expect(struct.label).to.equal('test');
+      expect(struct.valueSet.restricted).to.equal(true);
+      expect(struct.valueSet.valueSetDefinition.sorted === 'false').to.equal(false);
+      expect(struct.valueSet.valueSetDefinition.value[0].label).to.equal('value1');
+    });
+    it('should return Phone structure', () => {
+      struct = templates.createDefaultTypeStructure('Phone', 'Phone', 'test', null);
+      expect(struct.fullName).to.equal('Phone');
+      expect(struct.type).to.equal('Phone');
+      expect(struct.label).to.equal('test');
+    });
+    it('should return Text structure', () => {
+      struct = templates.createDefaultTypeStructure('Text', 'Text', 'test', null);
+      expect(struct.fullName).to.equal('Text');
+      expect(struct.type).to.equal('Text');
+      expect(struct.label).to.equal('test');
+      expect(struct.unique).to.equal(false);
+      expect(struct.length).to.equal(100);
+    });
+    it('should return TextArea structure', () => {
+      struct = templates.createDefaultTypeStructure('TextArea', 'TextArea', 'test', null);
+      expect(struct.fullName).to.equal('TextArea');
+      expect(struct.type).to.equal('TextArea');
+      expect(struct.label).to.equal('test');
+    });
+    it('should return LongTextArea structure', () => {
+      struct = templates.createDefaultTypeStructure('LongTextArea', 'LongTextArea', 'test', null);
+      expect(struct.fullName).to.equal('LongTextArea');
+      expect(struct.type).to.equal('LongTextArea');
+      expect(struct.label).to.equal('test');
+      expect(struct.length).to.equal(32768);
+      expect(struct.visibleLines).to.equal(3);
+    });
+    it('should return Url structure', () => {
+      struct = templates.createDefaultTypeStructure('Url', 'Url', 'test', null);
+      expect(struct.fullName).to.equal('Url');
+      expect(struct.type).to.equal('Url');
+      expect(struct.label).to.equal('test');
+    });
+    it('should return Default structure', () => {
+      struct = templates.createDefaultTypeStructure('DefaultA', 'DefaultB', 'DefaultC', null);
+      expect(struct.fullName).to.equal('DefaultA');
+      expect(struct.type).to.equal('DefaultB');
+      expect(struct.label).to.equal('DefaultC');
+    });
+  });
+  describe('canConvert', () => {
+    it('type should convert', () => {
+      [
+        'Checkbox',
+        'Date',
+        'DateTime',
+        'Email',
+        'Number',
+        'Percent',
+        'Phone',
+        'Picklist',
+        'Text',
+        'TextArea',
+        'LongTextArea',
+        'Url',
+      ].forEach((type) => {
+        expect(templates.canConvert(type)).to.be.true;
+      });
+    });
+    it('type should not convert', () => {
+      expect(templates.canConvert('test')).to.be.false;
+    });
+  });
 });
